requirements.txt
flask==3.1.0
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
Pillow>=9.0.0
easyocr>=1.6.0
readability-lxml>=0.9
gunicorn>=21.2.0
